backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1
mysql-connector-python==9.0.0
python-dotenv==1.0.1


